requirements.txt
anytree>=2.0.0,<3.0.0
beautifulsoup4>=4.0.0,<5.0.0
black>=24.0.0,<25.0.0
blinker>=1.0.0,<2.0.0
bs4>=0.0.1,<1.0.0
certifi>=2024.0.0,<2025.0.0
charset-normalizer>=3.0.0,<4.0.0
click>=8.0.0,<9.0.0
colorama>=0.0.4,<1.0.0
contourpy>=1.0.0,<2.0.0
cycler>=0.0.7,<1.0.0
DateTime>=5.0,<6.0
exceptiongroup>=1.0.0,<2.0.0
Flask>=3.0.0,<4.0.0
fonttools>=4.0.0,<5.0.0
idna>=3.0,<4.0
importlib_metadata>=8.0.0,<9.0.0
importlib-resources>=6.0.0,<7.0.0
iniconfig>=2.0.0,<3.0.0
itsdangerous>=2.0.0,<3.0.0
Jinja2>=3.0.0,<4.0.0
kiwisolver>=1.0.0,<2.0.0
logging>=0.0.1,<1.0.0
MarkupSafe>=2.0.0,<3.0.0
matplotlib>=3.0.0,<4.0.0
mypy>=1.0.0,<2.0.0
mypy-extensions>=1.0.0,<2.0.0
numpy>=2.0.0,<3.0.0
packaging>=24.0.0,<25.0.0
pandas~=2.2.2
pandas-stubs~=2.2.2.240909
pathspec>=0.12.1,<1.0.0
pillow>=10.0.0,<11.0.0
pip-review>=1.0.0,<2.0.0
platformdirs>=4.0.0,<5.0.0
pluggy>=1.5.0,<2.0.0
psutil>=6.0.0,<7.0.0
pylint>=3.0.0,<4.0.0
pyotp>=2.0.0,<3.0.0
pyparsing>=3.0.0,<4.0.0
pytest>=8.0.0,<9.0.0
python-dateutil>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
python-http-client>=3.0.0,<4.0.0
pytz>=2024.0.0,<2025.0.0
requests>=2.0.0,<3.0.0
sendgrid>=6.0.0,<7.0.0
setuptools>=69.0.0,<76.0.0
six>=1.0.0,<2.0.0
soupsieve>=2.0,<3.0
starkbank-ecdsa>=2.0.0,<3.0.0
tinydb>=4.0.0,<5.0.0
tomli~=2.0.1
tomlkit~=0.13.2
tqdm~=4.66.5
types-colorama~=0.4.15.20240311
types-psutil~=6.0.0.20240901
types-pytz~=2024.2.0.20240913
types-requests<=2.32.0.20240914
types-tqdm~=4.66.0.20240417
typing>=3.0.0.0,<4.0.0.0
typing_extensions>=4.0.0,<5.0.0
tzlocal>=5.0,<6.0
urllib3<3.0.0
Werkzeug>=3.0.0,<4.0.0
zipp>=3.0.0,<4.0.0
zope.interface>=6.0.0,<8.0.0
